moved to correct testfolder with Jira naming. fixed spacing issues
