+++ src/NHibernate.Test/NHSpecificTest/NH3984/Mappings.hbm.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <hibernate-mapping xmlns="urn:nhibernate-mapping-2.2"
                    assembly="NHibernate.Test"
-                   namespace="NHibernate.Test.NHSpecificTest.EntityWithDatetime2">
+                   namespace="NHibernate.Test.NHSpecificTest.NH3984">
 
   <class name="LogEntry">
     <id name="Id" type="Int32" generator="native" />
