+++ src/NHibernate/Loader/Custom/CustomLoader.cs
@@ -549,11 +549,11 @@ namespace NHibernate.Loader.Custom
 			/// <returns>The Hibernate type.</returns>
 			public IType GetHibernateType(int columnPos)
 			{
-                var type = TypeFactory.Basic(resultSet.GetDataTypeName(columnPos));
-                if (type == null)
-                    return TypeFactory.Basic(resultSet.GetFieldType(columnPos).Name);
-
-                return type;
+				var type = TypeFactory.Basic(resultSet.GetDataTypeName(columnPos));
+				if (type == null)
+					return TypeFactory.Basic(resultSet.GetFieldType(columnPos).Name);
+			
+				return type;
             }
 		}
 	}
